Get lists

Get lists

chmp_lists(
  conn,
  fields = NULL,
  exclude_fields = NULL,
  count = NULL,
  offset = NULL,
  before_date_created = NULL,
  since_date_created = NULL,
  before_campaign_last_sent = NULL,
  since_campaign_last_sent = NULL,
  email = NULL,
  sort_field = NULL,
  sort_dir = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

chmp_lists_(
  conn,
  fields = NULL,
  exclude_fields = NULL,
  count = NULL,
  offset = NULL,
  before_date_created = NULL,
  since_date_created = NULL,
  before_campaign_last_sent = NULL,
  since_campaign_last_sent = NULL,
  email = NULL,
  sort_field = NULL,
  sort_dir = NULL,
  key = NULL,
  ...
)

Arguments

conn
a connection object. see ChmpClient

fields
(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields
(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

count
(integer/numeric) The number of records to return. Default value is 10.

offset
(integer/numeric) The number of records from a collection to skip. Iterating over large collections with this parameter can be slow. Default value is 0.

before_date_created
(character) Restrict response to lists created before the set date.

since_date_created
(character) Restrict results to lists created after the set date.

before_campaign_last_sent
(character) Restrict results to lists created before the last campaign send date.

since_campaign_last_sent
(character) Restrict results to lists created after the last campaign send date.

email
(character) Restrict results to lists that include a specific subscriber's email address.

sort_field
(character) Returns files sorted by the specified field. Possible Values: date_created

sort_dir
(character) Determines the order direction for sorted results. Possible Values: ASC, DESC

key
A Mailchimp API key. See https://developer.mailchimp.com/documentation/mailchimp/ to get a key

parse
(logical) Whether to parse to list (FALSE) or data.frame (TRUE). Default: TRUE

...
Curl options passed to HttpClient

Examples

if (FALSE) {
(conn <- ChmpClient$new())
chmp_lists_(conn)
chmp_lists(conn)
chmp_lists(conn, parse = FALSE)
}