Lists class

Lists class

Lists class

Details

xxx

Note

not all routes are implemented thus far

Methods

Public methods

ChmpList$new()

ChmpList$print()

ChmpList$info()

ChmpList$activity()

ChmpList$abuse_reports()

ChmpList$clients()

ChmpList$growth_history()

ChmpList$interest_categories()

ChmpList$locations()

ChmpList$member()

ChmpList$member_activity()

ChmpList$member_goals()

ChmpList$member_notes()

ChmpList$members()

ChmpList$clone()

Method new()

Usage

ChmpList$new(conn, id)

Arguments

conn

a connection object. see ChmpClient

id

a mailchimp list id. required

Method print()

Usage

ChmpList$print(x, ...)

Method info()

Usage

ChmpList$info(
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method activity()

Usage

ChmpList$activity(
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method abuse_reports()

Usage

ChmpList$abuse_reports(
  report = NULL,
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method clients()

Usage

ChmpList$clients(
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method growth_history()

Usage

ChmpList$growth_history(
  month = NULL,
  fields = NULL,
  exclude_fields = NULL,
  count = 10,
  offset = 0,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

month

(character) a year/month character string of the year/month you want. optional

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

count

(integer/numeric) The number of records to return. Default value is 10.

offset

(integer/numeric) The number of records from a collection to skip. Iterating over large collections with this parameter can be slow. Default value is 0.

Method interest_categories()

Usage

ChmpList$interest_categories(
  category_id = NULL,
  fields = NULL,
  exclude_fields = NULL,
  count = 10,
  offset = 0,
  type = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

count

(integer/numeric) The number of records to return. Default value is 10.

offset

(integer/numeric) The number of records from a collection to skip. Iterating over large collections with this parameter can be slow. Default value is 0.

Method locations()

Usage

ChmpList$locations(
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method member()

Usage

ChmpList$member(
  subscriber_hash,
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method member_activity()

Usage

ChmpList$member_activity(
  subscriber_hash,
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method member_goals()

Usage

ChmpList$member_goals(
  subscriber_hash,
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method member_notes()

Usage

ChmpList$member_notes(
  subscriber_hash,
  fields = NULL,
  exclude_fields = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

Method members()

Usage

ChmpList$members(
  fields = NULL,
  exclude_fields = NULL,
  count = 10,
  offset = 0,
  email_type = NULL,
  status = NULL,
  since_timestamp_opt = NULL,
  before_timestamp_opt = NULL,
  since_last_changed = NULL,
  before_last_changed = NULL,
  unique_email_id = NULL,
  vip_only = NULL,
  interest_category_id = NULL,
  interest_ids = NULL,
  interest_match = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

Arguments

fields

(list/vector) A comma-separated list of fields to return. Reference parameters of sub-objects with dot notation.

exclude_fields

(list/vector) A comma-separated list of fields to exclude. Reference parameters of sub-objects with dot notation.

count

(integer/numeric) The number of records to return. Default value is 10.

offset

(integer/numeric) The number of records from a collection to skip. Iterating over large collections with this parameter can be slow. Default value is 0.

Method clone()

The objects of this class are cloneable with this method.

Usage

ChmpList$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
(conn <- ChmpClient$new())
(x <- ChmpList$new(conn, id = "<id>"))
x$info()
x$info(parse = FALSE)
x$info(fields = "stats")
x$info(fields = "stats.open_rate")
x$info(fields = "campaign_defaults")
x$activity()
x$abuse_reports()
x$clients()
x$growth_history()
x$growth_history("2018-02")
x$interest_categories()
x$locations()
x$members()
# x$member("<id>")
# x$member_activity("<id>")
# x$member_goals("<id>")
# x$member_notes("<id>")
}